+++ src/main/java/woowacourse/shoppingcart/domain/Quantity.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.domain.cartitem;
+package woowacourse.shoppingcart.domain;
 
 public class Quantity {
