+++ src/main/java/woowacourse/shoppingcart/dto/cartItem/CartItemResponse.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
 import com.fasterxml.jackson.annotation.JsonTypeInfo.Id;
 import com.fasterxml.jackson.annotation.JsonTypeName;
-import woowacourse.shoppingcart.domain.cartitem.CartItem;
+import woowacourse.shoppingcart.domain.CartItem;
 
 @JsonTypeName("cartItem")
 @JsonTypeInfo(include = As.WRAPPER_OBJECT, use = Id.NAME)
